Move over dateIsBlank
