+++ server/form-pages/apply/basic-information/oralHearing.ts
@@ -1,8 +1,7 @@
 import type { ObjectWithDateParts, YesOrNo } from 'approved-premises'
 
 import TasklistPage from '../../tasklistPage'
-import { dateIsBlank } from '../../../utils/utils'
-import { dateAndTimeInputsAreValidDates } from '../../../utils/dateUtils'
+import { dateAndTimeInputsAreValidDates, dateIsBlank } from '../../../utils/dateUtils'
 
 export default class OralHearing implements TasklistPage {
   name = 'oral-hearing'
