+++ server/utils/utils.test.ts
@@ -1,6 +1,5 @@
-import type { ObjectWithDateParts } from 'approved-premises'
 import { SessionDataError } from './errors'
-import { convertToTitleCase, initialiseName, retrieveQuestionResponseFromSession, dateIsBlank } from './utils'
+import { convertToTitleCase, initialiseName, retrieveQuestionResponseFromSession } from './utils'
 
 describe('convert to title case', () => {
   it.each([
@@ -44,27 +43,5 @@ describe('retrieveQuestionResponseFromSession', () => {
       'questionResponse',
     )
     expect(questionResponse).toBe('no')
-  })
-})
-
-describe('dateIsBlank', () => {
-  it('returns false if the date is not blank', () => {
-    const date: ObjectWithDateParts<'field'> = {
-      'field-day': '12',
-      'field-month': '1',
-      'field-year': '2022',
-    }
-
-    expect(dateIsBlank(date)).toEqual(false)
-  })
-
-  it('returns true if the date is blank', () => {
-    const date: ObjectWithDateParts<'field'> = {
-      'field-day': '',
-      'field-month': '',
-      'field-year': '',
-    }
-
-    expect(dateIsBlank(date)).toEqual(true)
   })
 })
